# data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "System Initiative API",
    "description": "The API Server for interacting with a System Initiative workspace",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "This API Server"
    }
  ],
  "paths": {
...
    "/v1/w/{workspace_id}/change-sets/{change_set_id}/components/{component_id}/action": {
      "post": {
        "tags": [
          "components"
        ],
        "operationId": "add_action",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "Workspace identifier",
            "required": true
          },
          {
            "name": "change_set_id",
            "in": "path",
            "description": "Change set identifier",
            "required": true
          },
          {
            "name": "component_id",
            "in": "path",
            "description": "Component identifier",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddActionV1Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Action successfully queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddActionV1Response"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "404": {
            "description": "Component or function not found"
          },
          "409": {
            "description": "action already enqueued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "412": {
            "description": "Precondition Failed - View not found or duplicate function name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation error - Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/w/{workspace_id}/change-sets/{change_set_id}/funcs/runs/{func_run_id}": {
      "get": {
        "tags": [
          "funcs"
        ],
        "operationId": "get_func_run",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "Workspace identifier",
            "required": true
          },
          {
            "name": "change_set_id",
            "in": "path",
            "description": "Change set identifier",
            "required": true
          },
          {
            "name": "func_run_id",
            "in": "path",
            "description": "Func run identifier",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Func Run retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetFuncRunV1Response"
                },
                "example": {
                  "funcRun": {
                    "id": "01JQCJ0AAXGX5M9QY10AVF4GK1",
                    "state": "Success",
                    "actor": "System",
                    "componentId": "01JP8KHZP3DZKGNXRP83Q6WTQ5",
                    "attributeValueId": null,
                    "componentName": "NAT Gateway IP 1",
                    "schemaName": "AWS::EC2::EIP",
                    "actionId": "01JQCHZZY99G3R0C1FA3W4AFR6",
                    "actionPrototypeId": "01JPNHEE9Z3DFW48XVZ1FX04KA",
                    "actionKind": "Destroy",
                    "actionDisplayName": "Destroy",
                    "actionOriginatingChangeSetId": "01JQCHZZVTAHHZ7DG0ZSCB9RXB",
                    "actionOriginatingChangeSetName": "2025-03-27-19:41",
                    "actionResultState": "Success",
                    "backendKind": "JsAction",
                    "backendResponseType": "Action",
                    "functionName": "Delete Asset",
                    "functionDisplayName": null,
                    "functionKind": "Action",
                    "functionDescription": null,
                    "functionLink": null,
                    "functionArgs": {
                      "properties": {
                        "domain": {
                          "Domain": "vpc",
                          "Tags": []
                        },
                        "resource": {
                          "payload": {
                            "AllocationId": "eipalloc-033720f9556a3b0c1",
                            "PublicIp": "3.213.242.163"
                          }
                        },
                        "si": {
                          "name": "NAT Gateway IP 1",
                          "resourceId": "3.213.242.163|eipalloc-033720f9556a3b0c1",
                          "type": "component"
                        }
                      }
                    },
                    "resultValue": {
                      "error": null,
                      "executionId": "01JQCJ0AAXGX5M9QY10AVF4GK1",
                      "message": null,
                      "payload": null,
                      "resourceId": null,
                      "status": "ok"
                    },
                    "logs": {
                      "id": "01JQCJ0ABJSCE01GNQDWVY1ZP5",
                      "createdAt": "2025-03-27T19:41:58.514416748Z",
                      "updatedAt": "2025-03-27T19:41:58.514416748Z",
                      "funcRunId": "01JQCJ0AAXGX5M9QY10AVF4GK1",
                      "logs": [
                        {
                          "stream": "stdout",
                          "executionId": "",
                          "level": "info",
                          "group": "log",
                          "message": "Running CLI command",
                          "timestamp": 1743104518
                        },
                        {
                          "stream": "output",
                          "executionId": "01JQCJ0AAXGX5M9QY10AVF4GK1",
                          "level": "info",
                          "group": "log",
                          "message": "Output: {\"status\":\"success\"}",
                          "timestamp": 1743104521
                        }
                      ],
                      "finalized": true
                    },
                    "createdAt": "2025-03-27T19:41:58.493298051Z",
                    "updatedAt": "2025-03-27T19:42:02.192033089Z"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "404": {
            "description": "Func run not found"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionReference": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "function"
            ],
            "properties": {
              "function": {
                "type": "string",
                "example": "Create Asset"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "actionPrototypeId"
            ],
            "properties": {
              "actionPrototypeId": {
                "type": "string",
                "example": "01H9ZQD35JPMBGHH69BT0Q79VY"
              }
            }
          }
        ],
        "description": "Reference to a management function by either name or ID.\nThis allows clients to use the more human-friendly name approach\nor the more precise ID approach when working with actions."
      },
...
      "AddActionV1Request": {
        "type": "object",
        "required": [
          "action"
        ],
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ActionReference"
          }
        },
        "example": {
          "action": {
            "function": "Create Asset"
          }
        }
      },
      "AddActionV1Response": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "example": {
          "success": true
        }
      },
      "ApiError": {
        "type": "object",
        "description": "Standard error response format for v1 API",
        "required": [
          "message",
          "statusCode"
        ],
        "properties": {
          "code": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "example": 4001
          },
          "message": {
            "type": "string",
            "example": "Invalid request data"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32",
            "example": 422,
            "minimum": 0
          }
        },
        "example": {
          "message": "Invalid request data",
          "statusCode": 422,
          "code": 4001
        }
      },
...
      "FuncRunLogViewV1": {
        "type": "object",
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "funcRunId",
          "logs",
          "finalized"
        ],
        "properties": {
          "createdAt": {
            "type": "string",
            "example": "2025-03-27T19:41:58.514416748Z"
          },
          "finalized": {
            "type": "boolean",
            "example": true
          },
          "funcRunId": {
            "type": "string",
            "example": "01JQCJ0AAXGX5M9QY10AVF4GK1"
          },
          "id": {
            "type": "string",
            "example": "01JQCJ0ABJSCE01GNQDWVY1ZP5"
          },
          "logs": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "example": [
              {
                "stream": "stdout",
                "executionId": "",
                "level": "info",
                "group": "log",
                "message": "Running CLI command: \"aws 'cloudcontrol' 'delete-resource'\"",
                "timestamp": 1743104518
              },
              {
                "stream": "output",
                "executionId": "01JQCJ0AAXGX5M9QY10AVF4GK1",
                "level": "info",
                "group": "log",
                "message": "Output: {\"protocol\":\"result\",\"status\":\"success\"}",
                "timestamp": 1743104521
              }
            ]
          },
          "updatedAt": {
            "type": "string",
            "example": "2025-03-27T19:41:58.514416748Z"
          }
        }
      },
...
      "FuncRunViewV1": {
        "type": "object",
        "required": [
          "id",
          "state",
          "attributeValueId",
          "componentId",
          "componentName",
          "schemaName",
          "actionId",
          "actionPrototypeId",
          "actionKind",
          "actionDisplayName",
          "actionOriginatingChangeSetId",
          "actionOriginatingChangeSetName",
          "actionResultState",
          "backendKind",
          "backendResponseType",
          "functionName",
          "functionDisplayName",
          "functionKind",
          "functionDescription",
          "functionLink",
          "functionArgs",
          "functionCodeBase64",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "actionDisplayName": {
            "type": "string",
            "example": "Destroy"
          },
          "actionId": {
            "type": "string",
            "example": "01JQCHZZY99G3R0C1FA3W4AFR6"
          },
          "actionKind": {
            "type": "string",
            "example": "Destroy"
          },
          "actionOriginatingChangeSetId": {
            "type": "string",
            "example": "01JQCHZZVTAHHZ7DG0ZSCB9RXB"
          },
          "actionOriginatingChangeSetName": {
            "type": "string",
            "example": "2025-03-27-19:41"
          },
          "actionPrototypeId": {
            "type": "string",
            "example": "01JPNHEE9Z3DFW48XVZ1FX04KA"
          },
          "actionResultState": {
            "type": "string",
            "example": "Success"
          },
          "attributeValueId": {
            "type": "string",
            "example": "null"
          },
          "backendKind": {
            "type": "string",
            "example": "JsAction"
          },
          "backendResponseType": {
            "type": "string",
            "example": "Action"
          },
          "componentId": {
            "type": "string",
            "example": "01JP8KHZP3DZKGNXRP83Q6WTQ5"
          },
          "componentName": {
            "type": "string",
            "example": "NAT Gateway IP 1"
          },
          "createdAt": {
            "type": "string",
            "example": "2025-03-27T19:41:58.493298051Z"
          },
          "functionArgs": {},
          "functionCodeBase64": {
            "type": "string",
            "example": "YXN5bmMgZnVuY3Rpb24gbWFpbihjb21wb2..."
          },
          "functionDescription": {
            "type": "string",
            "example": "null"
          },
          "functionDisplayName": {
            "type": "string",
            "example": "null"
          },
          "functionKind": {
            "type": "string",
            "example": "Action"
          },
          "functionLink": {
            "type": "string",
            "example": "null"
          },
          "functionName": {
            "type": "string",
            "example": "Delete Asset"
          },
          "id": {
            "type": "string",
            "example": "01JQCJ0AAXGX5M9QY10AVF4GK1"
          },
          "logs": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FuncRunLogViewV1"
              }
            ]
          },
          "resultValue": {},
          "schemaName": {
            "type": "string",
            "example": "AWS::EC2::EIP"
          },
          "state": {
            "type": "string",
            "example": "Success"
          },
          "updatedAt": {
            "type": "string",
            "example": "2025-03-27T19:42:02.192033089Z"
          }
        }
      },
...
      "GetFuncRunV1Response": {
        "type": "object",
        "required": [
          "funcRun"
        ],
        "properties": {
          "funcRun": {
            "$ref": "#/components/schemas/FuncRunViewV1"
          }
        }
      },
...
    }
  },
...
}